Add error message to syslog drain in case of dropped logs (#65)
* Initialise syslog agent with a non-null log client

The syslog connector is coded to emit an error log to its
log client when draining fails. However by default the syslog
connector is initialised with a null client which doesn't send
the logs anywhere.

This commit creates a loggregator ingress client and uses it when
creating the syslog connector in the syslog agent initialisation. This
results in the error log being sent to the firehose instead of
/dev/null.

* Add unit test for invalid GRPC certs in syslog agent config