Fix: golangci-lint warnings (#106)
* Ignore envstruct.WriteReport errors

If we've already loaded the env variables successfully, we shouldn't
error out on just printing the report. Maybe in the future we'd want to
log something if this function fails.

* Ignore gosec warning about pprof endpoints

We do put pprof endpoints on different ports, so we shouldn't be
receiving these warnings really.

* Fix deprecated gRPC functions

* Use LoggerV2 over deprecated Logger.
* Use insecure.Credentials over deprecated WithInsecure option.

* Ignore gosec where we need to set InsecureSkipVerify

* Use buffer.String() instead of string(buffer.Bytes())

* Resolve gosimple warnings

* Use `time.Since()` instead of `time.Now().Sub()`
* Remove unneccessary `fmt.Sprintf()` uses

* Add hel to tools

git.sr.ht/~nelsam/hel/v3 is a tool we use for generating mocks for
tests.

Adding hel to tools standardizes the version that we're using to
generate our mocks.

* Regenerate helheim mocks and remove unused mocks

* Fix the go generate tags in the packages.
* Regenerate mocks.
* Fix tests so they correctly use the new mocks; had to reduce
  concurrent data write test to only 40 iterations to fit within the new
  mock's buffer.
* Manually remove unused mocks.

* Fix ineffectual assignments

* Replace `time.Tick()`

`time.Tick()` only allows access to the underlying channel of the
ticker, meaning we can never stop the ticker. Even if it goes out of the
scope the goroutine will continue.

Since we want to stop the ticker if we exit this function, we should use
`time.NewTicker()` instead.

* Handle unchecked errors found by golangci-lint

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>
Signed-off-by: Duane May <mduane@vmware.com>

* Use the combined ListenAndServeTLS

handle errors once instead of on both calls

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>

* Remove unused code found by golangci-lint

Ignore the unused code generated by Helheim

Signed-off-by: Duane May <mduane@vmware.com>

* Ignore golangci-lint warnings about random number generation

golangci-lint gives security warnings when using the
random number generator from the math package instead
of the crypto package. In all of these cases, the math
package is fine, so we ignored the warnings.

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>

* Add specific IP to avoid binding to all network interfaces

Fixes gosec warnings from golangci-lint

Signed-off-by: Duane May <mduane@vmware.com>

* Update  WriteFile permissions to be 0600 or less

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>

* Increase TLS MinVersion

Signed-off-by: Duane May <mduane@vmware.com>

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>
Signed-off-by: Duane May <mduane@vmware.com>
Co-authored-by: Duane May <mduane@vmware.com>
Co-authored-by: Rebecca Roberts <robertsre@vmware.com>