Fix race condition in v1 conn manager test
Ensure that goroutine using mockConnector is closed before the `It`
block completes or else there will be a race condition on the
`mockConnector` variable from the outer `BeforeEach` block.

Put the channel write for `mockConnector.ConnectOutput.Ret2` into a case
so that it doesn't write if the channel is already full. When it was the
default case the goroutine could block there, preventing the goroutine
from closing for up to 60 seconds.