Add new protocol option to Forwarder Agent
Forwarder Agent parses a new field off the downstream destination
registration yaml files discovered in the file pattern provided. This new
field is `protocol`, which can either be left empty for the normal
behaviour, or filled in with `OTLP` to designate the destination as an
OTel Collector.

When the destination is designated as an OTel Collector, use the new
`otelcolclient` package to build a `Writer` to that Collector.