fix: rename go package in doppler.proto
Prepending `code.cloudfoundry.org` generates the code automatically into
`code.cloudfoundry.org/...`, which is not the path that we store the
loggregator-agent-release repo at typically. If we remove it, the
protobuf code generation step becomes a lot simpler.

Besides, we weren't even adding that line to our package name in the
generated files, so it's not very helpful. Better to have the path align
with where we want the generated files to actually go.