Legacy syslog bindings code cleanup
With the introduction of mTLS for Syslog Drains we have changed the
internal structure of the bindings. This PR cleans up the code of the
legacy syslog bindings and removes the http calls to the old Syslog
Binding Cache API endpoint. From now on, only the `/v2/binding` and
`/v2/aggregate-bindings` API endpoints of the Syslog Binding Cache will
be used.