Specify that counters are monotonic
Sums must be declared as monotonic in order to be represented in
prometheus as counters.

Note that this causes the otel collector to rename metrics when using
the prometheus exporter. Counters have `_total` appended to the name.

There are cases where counters aren't monotonic, in particular when
counters are emitted only as deltas (V1) and the counter totals are
reset in the Forwarder Agent.

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>