chore: Stabilize flaky loggr-forwarder-agent unit tests (#679)
Waits for the fake OTel log server to have data before purging any data, if log forwarding for OTel Collectors is turned on. This eliminates an edge case in the tests where data we meant to purge arrives in the fake OTel log server after we attempted to purge data from the server.