Fix Connection keepalive time for http batch
This change is necessary due to the current implementation details
of https_batch, having no retries.

Due to the common practice on many log drain targets, to enforce
a maximum keepalive time of usually 60s, this leads to
predictably failing requests after 60s for http_batch requests.

To prevent this issue, the setting
´MaxConnDuration` is necessary, to prevent connections to stay alive
longer than the servers keep-alive setting, leading to a
predestined rejection.